new util class
